Full month names, and a masthead that fits a 360px phone
"28 July" rather than "28 Jul", in the app, the Telegram mirror and the
push notification alike. This matters more in Hindi, where the short
form rendered as "जुल॰" — an abbreviation mark most readers would rather
not have to parse — and now reads "जुलाई".

The language switch becomes EN / हिं. Adding the theme button had pushed
the masthead past the width of a 360px screen; the full words were the
expensive part. Both keep their full names in aria-label and title, so
nothing is lost to a screen reader or a long-press, and the buttons keep
their 2.5rem touch target.

Verified at 360px: one row, and scrollWidth equal to the viewport.

+++ service/app/i18n.js
@@ -415,7 +415,7 @@ export function fmtWindow(fromIso, toIso, lang) {
   const locale = lang === 'hi' ? 'hi-IN' : 'en-IN';
   const from = new Date(fromIso);
   const to = new Date(toIso);
-  const day = new Intl.DateTimeFormat(locale, { weekday: 'short', day: 'numeric', month: 'short', timeZone: IST });
+  const day = new Intl.DateTimeFormat(locale, { weekday: 'short', day: 'numeric', month: 'long', timeZone: IST });
   const time = new Intl.DateTimeFormat(locale, { hour: 'numeric', minute: '2-digit', hour12: true, timeZone: IST });
   const sameDay = day.format(from) === day.format(to);
   const word = STRINGS[lang].to;
